ml/requirements.txt
scikit-learn>=1.4,<2.0
pandas>=2.0,<3.0
numpy>=1.26,<3.0
joblib>=1.3,<2.0
psycopg2-binary>=2.9,<3.0
httpx>=0.27,<1.0
